requirements.txt
easyocr>=1.7.0
pynput>=1.7.6
mss>=9.0.1
Pillow>=10.0.0
